Refactored routes, requests now don't need to pass the user id anymore

+++ src/main/kotlin/com/fifty/routes/PostRoutes.kt
@@ -3,7 +3,6 @@ package com.fifty.routes
 import com.fifty.data.requests.CreatePostRequest
 import com.fifty.data.requests.DeletePostRequest
 import com.fifty.data.responses.BasicApiResponse
-import com.fifty.plugins.email
 import com.fifty.service.LikeService
 import com.fifty.service.PostService
 import com.fifty.service.UserService
@@ -13,14 +12,12 @@ import com.fifty.util.QueryParams
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
-import io.ktor.server.auth.jwt.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
 fun Route.createPost(
     postService: PostService,
-    userService: UserService
 ) {
     authenticate {
         post("/api/post/create") {
@@ -29,28 +26,24 @@ fun Route.createPost(
                     call.respond(HttpStatusCode.BadRequest)
                     return@post
                 }
+            val userId = call.userId
 
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ) {
-                val didUserExist = postService.createPostIfUserExists(request)
-                if (!didUserExist) {
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = false,
-                            message = ApiResponseMessages.USER_NOT_FOUND
-                        )
+            val didUserExist = postService.createPostIfUserExists(request, userId)
+            if (!didUserExist) {
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = false,
+                        message = ApiResponseMessages.USER_NOT_FOUND
                     )
-                } else {
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = true
-                        )
+                )
+            } else {
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = true
                     )
-                }
+                )
             }
         }
     }
@@ -58,60 +51,49 @@ fun Route.createPost(
 
 fun Route.getPostsForFollows(
     postService: PostService,
-    userService: UserService
 ) {
     authenticate {
         get("/api/post/get") {
-            val userId = call.parameters[QueryParams.PARAM_USER_ID] ?: kotlin.run {
-                call.respond(HttpStatusCode.BadRequest)
-                return@get
-            }
             val page = call.parameters[QueryParams.PARAM_PAGE]?.toIntOrNull() ?: 0
             val pageSize =
                 call.parameters[QueryParams.PARAM_PAGE_SIZE]?.toIntOrNull() ?: Constants.DEFAULT_POST_PAGE_SIZE
 
-            ifEmailBelongsToUser(
-                userId = userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ) {
-                val posts = postService.getPostsForFollows(
-                    userId, page, pageSize
-                )
-                call.respond(
-                    HttpStatusCode.OK,
-                    posts
-                )
-            }
+            val posts = postService.getPostsForFollows(
+                call.userId, page, pageSize
+            )
+            call.respond(
+                HttpStatusCode.OK,
+                posts
+            )
         }
     }
 }
 
 fun Route.deletePost(
     postService: PostService,
-    userService: UserService,
     likeService: LikeService
 ) {
-    delete("/api/post/delete") {
-        val request =
-            kotlin.runCatching { call.receiveNullable<DeletePostRequest>() }.getOrNull() ?: kotlin.run {
-                call.respond(HttpStatusCode.BadRequest)
+    authenticate {
+        delete("/api/post/delete") {
+            val request =
+                kotlin.runCatching { call.receiveNullable<DeletePostRequest>() }.getOrNull() ?: kotlin.run {
+                    call.respond(HttpStatusCode.BadRequest)
+                    return@delete
+                }
+
+            val post = postService.getPost(request.postId)
+            if (post == null) {
+                call.respond(HttpStatusCode.NotFound)
                 return@delete
             }
-
-        val post = postService.getPost(request.postId)
-        if (post == null) {
-            call.respond(HttpStatusCode.NotFound)
-            return@delete
-        }
-
-        ifEmailBelongsToUser(
-            userId = post.userId,
-            validateEmail = userService::doesEmailBelongToUserId
-        ) {
-            postService.deletePost(postId = request.postId)
-            likeService.deleteLikesForParent(request.postId)
-            // TODO: Delete comments from post
-            call.respond(HttpStatusCode.OK)
+            if (post.userId == call.userId) {
+                postService.deletePost(postId = request.postId)
+                likeService.deleteLikesForParent(request.postId)
+                // TODO: Delete comments from post
+                call.respond(HttpStatusCode.OK)
+            } else {
+                call.respond(HttpStatusCode.Unauthorized)
+            }
         }
     }
 }
